+++ src/cytomine.js
@@ -1,14 +1,28 @@
 import axios from 'axios';
 
+function prepareBasePath(basePath) {
+  if (!basePath.startsWith('/')) {
+    basePath = '/' + basePath;
+  }
+
+  if (!basePath.endsWith('/')) {
+    basePath = basePath + '/';
+  }
+
+  return basePath;
+}
+
 export default class Cytomine {
 
   /**
    * @param {string}   host             The Cytomine host
    * @param {string} [basePath=/api/]   The base path to perform API requests
+   * @param {string} iamPath            The base path to perform IAM requests
+   * @param {function} authorizationHeaderInterceptor   The Axios request interceptor for authorization header
    *
    * @returns {this} The singleton instance
    */
-  constructor(host, basePath = '/api/') {
+  constructor(host, basePath = '/api/', iamPath = '/iam/realms/cytomine/', authorizationHeaderInterceptor = null) {
     if(!Cytomine._instance) {
       if(!host.startsWith('http://') && !host.startsWith('https://')) {
         host = 'http://' + host;
@@ -17,109 +31,39 @@ export default class Cytomine {
         host = host.slice(0, -1);
       }
 
-      if(!basePath.startsWith('/')) {
-        basePath = '/' + basePath;
-      }
-
-      if(!basePath.endsWith('/')) {
-        basePath = basePath + '/';
-      }
-
       this._host = host;
-      this._basePath = basePath;
+      this._basePath = prepareBasePath(basePath);
+      this._iamPath = prepareBasePath(iamPath);
+
+      const onRejectedResponseInterceptor = error => {
+        error.message += ' - Response data: ' + JSON.stringify(error.response.data);
+        return Promise.reject(error);
+      };
+
+      this.iam = axios.create({
+        baseURL: this._host + this._iamPath,
+        withCredentials: true
+      });
+      if (authorizationHeaderInterceptor !== null) {
+        this.iam.interceptors.request.use(authorizationHeaderInterceptor);
+      }
+      this.iam.interceptors.response.use(response => response, onRejectedResponseInterceptor);
 
       this.api = axios.create({
         baseURL: this._host + this._basePath,
         withCredentials: true
       });
-
-      this.api.interceptors.response.use(function (response) {
-        return response;
-      }, function (error) {
-        error.message += ' - Response data: ' + JSON.stringify(error.response.data);
-        return Promise.reject(error);
-      });
-
-      let readToken = function() {
-        let switchedToken = sessionStorage.getItem('cytomine-switched-user-token');
-
-        if (switchedToken!=null) {
-          let validity = sessionStorage.getItem('cytomine-switched-user-validity');
-          let diffTime = sessionStorage.getItem('cytomine-switched-user-server-diff-time');
-          let serverDate = new Date(new Date().getTime() - diffTime);
-          if (serverDate > validity) {
-            // sessionStorage.removeItem('cytomine-switched-user-token');
-            // sessionStorage.removeItem('cytomine-switched-user-server-diff-time');
-            return localStorage.getItem('cytomine-authentication-token') || sessionStorage.getItem('cytomine-authentication-token');
-          }
-          else {
-            return sessionStorage.getItem('cytomine-switched-user-token');
-          }
-        }
-        else {
-          return localStorage.getItem('cytomine-authentication-token') || sessionStorage.getItem('cytomine-authentication-token');
-        }
-      };
-
-
-      const onRequestSuccess = config => {
-        const token = readToken();
-        if (token) {
-          if (!config.headers) {
-            config.headers = {};
-          }
-          config.headers.Authorization = `Bearer ${token}`;
-        }
-        return config;
-      };
-      this.api.interceptors.request.use(onRequestSuccess);
-
-
+      if (authorizationHeaderInterceptor !== null) {
+        this.api.interceptors.request.use(authorizationHeaderInterceptor);
+      }
+      this.api.interceptors.response.use(response => response,  onRejectedResponseInterceptor);
       this.lastCommand = null;
 
       Cytomine._instance = this;
     }
 
-
-    let readToken = function() {
-      let switchedToken = sessionStorage.getItem('cytomine-switched-user-token');
-
-      if (switchedToken!=null) {
-        let validity = sessionStorage.getItem('cytomine-switched-user-validity');
-        let diffTime = sessionStorage.getItem('cytomine-switched-user-server-diff-time');
-        let serverDate = new Date(new Date().getTime() + diffTime);
-        if (serverDate > validity) {
-          // sessionStorage.removeItem('cytomine-switched-user-token');
-          // sessionStorage.removeItem('cytomine-switched-user-server-diff-time');
-          return localStorage.getItem('cytomine-authentication-token') || sessionStorage.getItem('cytomine-authentication-token');
-        }
-        else {
-          return sessionStorage.getItem('cytomine-switched-user-token');
-        }
-      }
-      else {
-        return localStorage.getItem('cytomine-authentication-token') || sessionStorage.getItem('cytomine-authentication-token');
-      }
-    };
-
-    const onRequestSuccess = config => {
-      const token = readToken();
-      if (token) {
-        if (!config.headers) {
-          config.headers = {};
-        }
-        config.headers.Authorization = `Bearer ${token}`;
-      }
-      return config;
-    };
-    axios.interceptors.request.use(onRequestSuccess);
-
-
     return Cytomine._instance;
   }
-
-
-
 
   /**
    * @returns {this} The singleton instance
@@ -152,125 +96,9 @@ export default class Cytomine {
    * @returns {{alive, authenticated, version, serverURL, serverID, user}} The data returned by the server
    */
   async ping(project) {
-
-    //
-    // const token = localStorage.getItem('cytomine-authentication-token') || sessionStorage.getItem('cytomine-authentication-token');
-    //
-    // let tokenValue = 'unknown';
-    // if (token) {
-    //   tokenValue = `Bearer ${token}`;
-    // }
-    // let {data} = await axios.post(`${this._host}/server/ping.json`, {project}, {headers: {'Authorization': `${tokenValue}`}});
-
-    let {data} = await axios.post(`${this._host}/server/ping.json`, {project}, {withCredentials: true});
+    let {data} = await this.api.post(`${this._host}/server/ping.json`, {project});
 
     return data;
-  }
-
-  /**
-   * Login to Cytomine with the provided credentials
-   *
-   * @param {string}  username            The username
-   * @param {string}  password            The password
-   * @param {boolean} [rememberMe=true]   Whether or not to remember the user
-   */
-  async login(username, password, rememberMe=true) {
-    const params = { username: username, password: password, rememberMe: rememberMe };
-    let {data} = await axios.post(`${this._host}/api/authenticate`, params, {withCredentials: true});
-    const jwt = data['token'];
-
-    if (rememberMe) {
-      localStorage.setItem('cytomine-authentication-token', jwt);
-      sessionStorage.removeItem('cytomine-authentication-token');
-    }
-    else {
-      sessionStorage.setItem('cytomine-authentication-token', jwt);
-      localStorage.removeItem('cytomine-authentication-token');
-    }
-
-    return data;
-  }
-
-  /**
-   * Login to Cytomine with a token
-   *
-   * @param {string}  username    The username
-   * @param {string}  tokenKey    The token
-   */
-  async loginWithToken(username, tokenKey) {
-    let params = {username, tokenKey};
-    let result = await axios.get(`${this._host}/login/loginWithToken`, {withCredentials: true, params});
-    const jwt = result.data['token'];
-
-    sessionStorage.setItem('cytomine-authentication-token', jwt);
-    localStorage.removeItem('cytomine-authentication-token');
-
-    return result.data;
-  }
-
-
-  async token(username, validity) {
-    let params = {username, validity};
-    let result = await axios.get(`${this._host}/api/token.json`, {withCredentials: true, params});
-    const token = result.data['token'];
-    return token;
-  }
-
-  /**
-   * Impersonate another user
-   *
-   * @param {String} username The username of the user to impersonate
-   */
-  async switchUser(username) {
-    let params = new URLSearchParams();
-    params.append('username', username);
-    let result = await axios.post(`${this._host}/api/login/impersonate`, params, {withCredentials: true});
-    sessionStorage.setItem('cytomine-switched-user-token', result.data['id_token']);
-    sessionStorage.setItem('cytomine-switched-user-validity', result.data['validity']);
-    sessionStorage.setItem('cytomine-switched-user-server-diff-time', result.data['created']-new Date().getTime());
-
-  }
-
-  /**
-   * Stops impersonating another user
-   */
-  stopSwitchUser() {
-    sessionStorage.removeItem('cytomine-switched-user-token');
-    sessionStorage.removeItem('cytomine-switched-user-validity');
-    sessionStorage.removeItem('cytomine-switched-user-created');
-  }
-
-  /**
-   * Logout from Cytomine
-   */
-  logout() {
-    sessionStorage.removeItem('cytomine-authentication-token');
-    localStorage.removeItem('cytomine-authentication-token');
-    sessionStorage.removeItem('cytomine-switched-user-token');
-    sessionStorage.removeItem('cytomine-switched-user-validity');
-    sessionStorage.removeItem('cytomine-switched-user-created');
-  }
-
-  /**
-   * Send a mail to the provided email address with the associated username
-   *
-   * @param {string}  email   The email address
-   */
-  async forgotUsername(email) {
-    let params = new URLSearchParams();
-    params.append('j_email', email);
-    await axios.post(`${this._host}/login/forgotUsername`, params);
-  }
-
-  /**
-   * Send a mail to the email associated to the provided username to reset password
-   *
-   * @param {string}  username    The username
-   */
-  async forgotPassword(username) {
-    let params = new URLSearchParams();
-    params.append('j_username', username);
-    await axios.post(`${this._host}/login/forgotPassword`, params);
   }
 
   /**
@@ -299,11 +127,11 @@ export default class Cytomine {
    */
   async fetchUIConfigCurrentUser(project) {
     let params = {};
-    if(project) {
+    if (project) {
       params.project = project;
     }
 
-    let {data} = await axios.get(`${this._host}/custom-ui/config.json`, {params, withCredentials: true});
+    let {data} = await this.api.get(`${this._host}/custom-ui/config.json`, {params});
     return data;
   }