+++ src/models/project.js
@@ -3,7 +3,6 @@ import Model from './model.js';
 import User from './user.js';
 import UserCollection from '../collections/user-collection.js';
 import {AnnotationType} from './annotation.js';
-import axios from 'axios';
 
 /** Enum providing the project member roles handled in Cytomine */
 export const ProjectMemberRole = Object.freeze({
@@ -263,8 +262,7 @@ export default class Project extends Model {
       throw new Error('Cannot fetch UI configuration of a project with no ID.');
     }
 
-    let {data} = await axios.get(`${Cytomine.instance._host}/custom-ui/project/${this.id}.json`,
-      {withCredentials: true});
+    let {data} = await this.api.get(`${Cytomine.instance._host}/custom-ui/project/${this.id}.json`);
     return data;
   }
 
@@ -279,8 +277,7 @@ export default class Project extends Model {
       throw new Error('Cannot save UI configuration of a project with no ID.');
     }
 
-    let {data} = await axios.post(`${Cytomine.instance.host}/custom-ui/project/${this.id}.json`, config,
-      {withCredentials: true});
+    let {data} = await this.api.post(`${Cytomine.instance.host}/custom-ui/project/${this.id}.json`, config);
     return data;
   }